Revamped wallet logic

+++ quart_app/maindir/indyutils/agent.py
@@ -1,18 +1,20 @@
-import asyncio
 import json
 import logging
 import os
-from indy import pool, ledger, wallet, did, crypto
-from indy.error import IndyError
 
+from indy import pool
+from indy.error import IndyError, ErrorCode
+
+from ..indyutils.utils import open_close_wallet
 from ..indyutils.wallet.agentwallet import AgentWallet
-from ..indyutils.utils import open_close_pool
+
+log = logging.getLogger(__name__)
 
 
 class Agent(AgentWallet):
-    log = logging.getLogger("indydev")
 
     def __init__(self):
+        super().__init__()
         self.genesis_file_path = os.path.abspath(
             os.path.join(os.path.dirname(__file__), "..", "ptgenesis")
         )
@@ -32,9 +34,11 @@ class Agent(AgentWallet):
             genesis_file_path=self.genesis_file_path,
             version=self.PROTOCOL_VERSION,
         )
-
-    async def get_agent_did(self):
-        return await self.didfromseed(seed=self.agent_seed)
+    @open_close_wallet
+    async def get_agent_did(self, **kwargs):
+        agent_did = await self.didfromseed(seed=self.agent_seed, wallet_handle=kwargs["wallet_handle"])
+        log.debug(f"Agent DID: {agent_did}")
+        return agent_did
 
     async def getverkey(self):
         return await self.verkeyfromseed(seed=self.agent_seed)
@@ -48,28 +52,32 @@ class Agent(AgentWallet):
         pool_config = json.dumps({"genesis_txn": genesis_file_path})
         try:
             await pool.create_pool_ledger_config(pool_name, pool_config)
-        except IndyError:
-            await pool.delete_pool_ledger_config(config_name=pool_name)
-            await pool.create_pool_ledger_config(pool_name, pool_config)
+        except IndyError as e:
+            if e.error_code == ErrorCode.PoolLedgerConfigAlreadyExistsError:
+                log.warning("Pool config already created!")
+                pass
+            else:
+                await pool.delete_pool_ledger_config(config_name=pool_name)
+                await pool.create_pool_ledger_config(pool_name, pool_config)
 
-    @open_close_pool
-    async def submitToPool(
-        self,
-        target_did=None,
-        target_ver_key=None,
-        alias=None,
-        pool_handle=None,
-        wallet_handle=None,
-    ):
-        agent_did = await self.get_agent_did()
-
-        nym_request_json = await self._build_nym_request(
-            agent_did, target_did, target_ver_key, alias=alias, role=self.agent_role
-        )
-
-        await ledger.sign_and_submit_request(
-            pool_handle=pool_handle,
-            wallet_handle=wallet_handle,
-            submitter_did=agent_did,
-            request_json=nym_request_json,
-        )
+    # @open_close_pool
+    # async def submitToPool(
+    #     self,
+    #     target_did=None,
+    #     target_ver_key=None,
+    #     alias=None,
+    #     pool_handle=None,
+    #     wallet_handle=None,
+    # ):
+    #     agent_did = await self.get_agent_did()
+    #
+    #     nym_request_json = await self._build_nym_request(
+    #         agent_did, target_did, target_ver_key, alias=alias, role=self.agent_role
+    #     )
+    #
+    #     await ledger.sign_and_submit_request(
+    #         pool_handle=pool_handle,
+    #         wallet_handle=wallet_handle,
+    #         submitter_did=agent_did,
+    #         request_json=nym_request_json,
+    #     )
